Restructure packages following DDD and hexagonal architecture principles. Refactored logging and removed duplicate classes.
