--- ${LOG_PATH}/app.json
@@ -1,1 +1,0 @@
-[{"instant":{"epochSecond":1756245794,"nanoOfSecond":97426500},"thread":"Test worker","level":"WARN","loggerName":"org.hibernate.orm.deprecation","message":"HHH90000025: MySQL8Dialect does not need to be specified explicitly using 'hibernate.dialect' (remove the property setting and it will be selected by default)","endOfBatch":false,"loggerFqcn":"org.hibernate.internal.log.DeprecationLogger_$logger","threadId":1,"threadPriority":5,"traceId":"${ctx:traceId}","spanId":"${ctx:spanId}"}, {"instant":{"epochSecond":1756245794,"nanoOfSecond":219280100},"thread":"Test worker","level":"WARN","loggerName":"org.hibernate.orm.deprecation","message":"HHH90000026: MySQL8Dialect has been deprecated; use org.hibernate.dialect.MySQLDialect instead","endOfBatch":false,"loggerFqcn":"org.hibernate.internal.log.DeprecationLogger_$logger","threadId":1,"threadPriority":5,"traceId":"${ctx:traceId}","spanId":"${ctx:spanId}"}, {"instant":{"epochSecond":1756245797,"nanoOfSecond":39187100},"thread":"Test worker","level":"WARN","loggerName":"org.springframework.boot.autoconfigure.orm.jpa.JpaBaseConfiguration$JpaWebConfiguration","message":"spring.jpa.open-in-view is enabled by default. Therefore, database queries may be performed during view rendering. Explicitly configure spring.jpa.open-in-view to disable this warning","endOfBatch":false,"loggerFqcn":"org.apache.commons.logging.LogAdapter$Log4jLog","threadId":1,"threadPriority":5,"traceId":"${ctx:traceId}","spanId":"${ctx:spanId}"}, {"instant":{"epochSecond":1756245797,"nanoOfSecond":432336500},"thread":"Test worker","level":"WARN","loggerName":"org.springframework.security.config.annotation.web.builders.WebSecurity","message":"\n\n********************************************************************\n**********        Security debugging is enabled.       *************\n**********    This may include sensitive information.  *************\n**********      Do not use in a production system!     *************\n********************************************************************\n\n","endOfBatch":false,"loggerFqcn":"org.apache.commons.logging.LogAdapter$Log4jLog","threadId":1,"threadPriority":5,"traceId":"${ctx:traceId}","spanId":"${ctx:spanId}"}]{"instant":{"epochSecond":1756245878,"nanoOfSecond":889397000},"thread":"restartedMain","level":"WARN","loggerName":"org.hibernate.orm.deprecation","message":"HHH90000025: MySQL8Dialect does not need to be specified explicitly using 'hibernate.dialect' (remove the property setting and it will be selected by default)","endOfBatch":false,"loggerFqcn":"org.hibernate.internal.log.DeprecationLogger_$logger","threadId":30,"threadPriority":5,"traceId":"${ctx:traceId}","spanId":"${ctx:spanId}"}, {"instant":{"epochSecond":1756245878,"nanoOfSecond":970755000},"thread":"restartedMain","level":"WARN","loggerName":"org.hibernate.orm.deprecation","message":"HHH90000026: MySQL8Dialect has been deprecated; use org.hibernate.dialect.MySQLDialect instead","endOfBatch":false,"loggerFqcn":"org.hibernate.internal.log.DeprecationLogger_$logger","threadId":30,"threadPriority":5,"traceId":"${ctx:traceId}","spanId":"${ctx:spanId}"}, {"instant":{"epochSecond":1756245881,"nanoOfSecond":824384500},"thread":"restartedMain","level":"WARN","loggerName":"org.springframework.boot.autoconfigure.orm.jpa.JpaBaseConfiguration$JpaWebConfiguration","message":"spring.jpa.open-in-view is enabled by default. Therefore, database queries may be performed during view rendering. Explicitly configure spring.jpa.open-in-view to disable this warning","endOfBatch":false,"loggerFqcn":"org.apache.commons.logging.LogAdapter$Log4jLog","threadId":30,"threadPriority":5,"traceId":"${ctx:traceId}","spanId":"${ctx:spanId}"}, {"instant":{"epochSecond":1756245882,"nanoOfSecond":232557300},"thread":"restartedMain","level":"WARN","loggerName":"org.springframework.security.config.annotation.web.builders.WebSecurity","message":"\n\n********************************************************************\n**********        Security debugging is enabled.       *************\n**********    This may include sensitive information.  *************\n**********      Do not use in a production system!     *************\n********************************************************************\n\n","endOfBatch":false,"loggerFqcn":"org.apache.commons.logging.LogAdapter$Log4jLog","threadId":30,"threadPriority":5,"traceId":"${ctx:traceId}","spanId":"${ctx:spanId}"}{"instant":{"epochSecond":1756247740,"nanoOfSecond":661266300},"thread":"Test worker","level":"WARN","loggerName":"org.hibernate.orm.deprecation","message":"HHH90000025: MySQL8Dialect does not need to be specified explicitly using 'hibernate.dialect' (remove the property setting and it will be selected by default)","endOfBatch":false,"loggerFqcn":"org.hibernate.internal.log.DeprecationLogger_$logger","threadId":1,"threadPriority":5,"traceId":"${ctx:traceId}","spanId":"${ctx:spanId}"}, {"instant":{"epochSecond":1756247740,"nanoOfSecond":739404100},"thread":"Test worker","level":"WARN","loggerName":"org.hibernate.orm.deprecation","message":"HHH90000026: MySQL8Dialect has been deprecated; use org.hibernate.dialect.MySQLDialect instead","endOfBatch":false,"loggerFqcn":"org.hibernate.internal.log.DeprecationLogger_$logger","threadId":1,"threadPriority":5,"traceId":"${ctx:traceId}","spanId":"${ctx:spanId}"}, {"instant":{"epochSecond":1756247743,"nanoOfSecond":523643300},"thread":"Test worker","level":"WARN","loggerName":"org.springframework.boot.autoconfigure.orm.jpa.JpaBaseConfiguration$JpaWebConfiguration","message":"spring.jpa.open-in-view is enabled by default. Therefore, database queries may be performed during view rendering. Explicitly configure spring.jpa.open-in-view to disable this warning","endOfBatch":false,"loggerFqcn":"org.apache.commons.logging.LogAdapter$Log4jLog","threadId":1,"threadPriority":5,"traceId":"${ctx:traceId}","spanId":"${ctx:spanId}"}]{"instant":{"epochSecond":1756247835,"nanoOfSecond":998011600},"thread":"restartedMain","level":"WARN","loggerName":"org.hibernate.orm.deprecation","message":"HHH90000025: MySQL8Dialect does not need to be specified explicitly using 'hibernate.dialect' (remove the property setting and it will be selected by default)","endOfBatch":false,"loggerFqcn":"org.hibernate.internal.log.DeprecationLogger_$logger","threadId":30,"threadPriority":5,"traceId":"${ctx:traceId}","spanId":"${ctx:spanId}"}, {"instant":{"epochSecond":1756247836,"nanoOfSecond":82327500},"thread":"restartedMain","level":"WARN","loggerName":"org.hibernate.orm.deprecation","message":"HHH90000026: MySQL8Dialect has been deprecated; use org.hibernate.dialect.MySQLDialect instead","endOfBatch":false,"loggerFqcn":"org.hibernate.internal.log.DeprecationLogger_$logger","threadId":30,"threadPriority":5,"traceId":"${ctx:traceId}","spanId":"${ctx:spanId}"}, {"instant":{"epochSecond":1756247838,"nanoOfSecond":784846100},"thread":"restartedMain","level":"WARN","loggerName":"org.springframework.boot.autoconfigure.orm.jpa.JpaBaseConfiguration$JpaWebConfiguration","message":"spring.jpa.open-in-view is enabled by default. Therefore, database queries may be performed during view rendering. Explicitly configure spring.jpa.open-in-view to disable this warning","endOfBatch":false,"loggerFqcn":"org.apache.commons.logging.LogAdapter$Log4jLog","threadId":30,"threadPriority":5,"traceId":"${ctx:traceId}","spanId":"${ctx:spanId}"}
