+++ src/main/java/com/propydis/studio/config/DotenvConfig.java
@@ -1,0 +1,24 @@
+package com.propydis.studio.config;
+
+import io.github.cdimascio.dotenv.Dotenv;
+import org.springframework.context.annotation.Configuration;
+
+import jakarta.annotation.PostConstruct;
+
+@Configuration
+public class DotenvConfig {
+
+    @PostConstruct
+    public void loadEnv() {
+        Dotenv dotenv = Dotenv.configure()
+                .filename(".env")
+                .ignoreIfMissing()
+                .load();
+
+        dotenv.entries().forEach(entry ->
+                System.setProperty(entry.getKey(), entry.getValue())
+        );
+    }
+}
+
+
