--- src/main/resources/log4j2.xml
@@ -1,63 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<Configuration status="WARN" monitorInterval="30">
-
-    <Properties>
-        <Property name="LOG_PATH">${spring:logging.file.path:-logs}</Property>
-        <Property name="LOG_LEVEL">${spring:logging.level.root:-WARN}</Property>
-    </Properties>
-
-    <!-- Environment configuration 'dev' -->
-    <SpringProfile name="dev">
-        <Properties>
-            <Property name="DEV_LOG_PATH">logs</Property>
-            <Property name="DEV_LOG_LEVEL">DEBUG</Property>
-        </Properties>
-        <Appenders>
-            <Console name="ConsoleDev" target="SYSTEM_OUT">
-                <PatternLayout pattern="%d{HH:mm:ss.SSS} [%t] %-5level %logger{36} - %msg%n"/>
-            </Console>
-            <RollingFile name="DevFile"
-                         fileName="${DEV_LOG_PATH}/dev-app.json"
-                         filePattern="${DEV_LOG_PATH}/archive/app-%d{yyyy-MM-dd}-%i.json.gz">
-
-                <JSONLayout complete="false" compact="true" eventEol="true"/>
-                <Policies>
-                    <TimeBasedTriggeringPolicy interval="1" modulate="true"/>
-                    <SizeBasedTriggeringPolicy size="50 MB"/>
-                </Policies>
-            </RollingFile>
-        </Appenders>
-        <Loggers>
-            <Root level="${DEV_LOG_LEVEL}">
-                <AppenderRef ref="ConsoleDev"/>
-                <AppenderRef ref="DevFile"/>
-            </Root>
-        </Loggers>
-    </SpringProfile>
-
-    <!-- Environment configuration 'prod' -->
-    <SpringProfile name="prod">
-        <Appenders>
-            <Console name="ConsoleProd" target="SYSTEM_OUT">
-                <JSONLayout complete="false" compact="true"/>
-            </Console>
-            <RollingFile name="ProdFile"
-                         fileName="${LOG_PATH}/app.json"
-                         filePattern="${LOG_PATH}/archive/app-%d{yyyy-MM-dd}-%i.json.gz">
-
-                <JSONLayout complete="false" compact="true" eventEol="true"/>
-                <Policies>
-                    <TimeBasedTriggeringPolicy interval="1" modulate="true"/>
-                    <SizeBasedTriggeringPolicy size="100 MB"/>
-                </Policies>
-            </RollingFile>
-        </Appenders>
-        <Loggers>
-            <Root level="${LOG_LEVEL}">
-                <AppenderRef ref="ConsoleProd"/>
-                <AppenderRef ref="ProdFile"/>
-            </Root>
-        </Loggers>
-    </SpringProfile>
-
-</Configuration>
